// src/es/schemas/_types.aggregations.ts
export const AggregationsMissing = z.union([z.string(), z.lazy(() => integer), z.lazy(() => double), z.boolean()]).meta({ id: 'AggregationsMissing' })
export type AggregationsMissing = z.infer<typeof AggregationsMissing>
// ...
export const AggregationsCalendarInterval = z.enum(['second', '1s', 'minute', '1m', 'hour', '1h', 'day', '1d', 'week', '1w', 'month', '1M', 'quarter', '1q', 'year', '1y']).meta({ id: 'AggregationsCalendarInterval' })
export type AggregationsCalendarInterval = z.infer<typeof AggregationsCalendarInterval>

export const AggregationsExtendedBounds = z.object({
  max: z.any().describe('Maximum value for the bound.').optional(),
  min: z.any().describe('Minimum value for the bound.').optional()
}).meta({ id: 'AggregationsExtendedBounds' })
export type AggregationsExtendedBounds = z.infer<typeof AggregationsExtendedBounds>

export const AggregationsAggregateOrder = z.union([z.record(z.lazy(() => Field), z.lazy(() => SortOrder)), z.array(z.record(z.lazy(() => Field), z.lazy(() => SortOrder)))]).meta({ id: 'AggregationsAggregateOrder' })
export type AggregationsAggregateOrder = z.infer<typeof AggregationsAggregateOrder>
// ...
export const AggregationsDateHistogramAggregation = z.object({
  calendar_interval: AggregationsCalendarInterval.describe('Calendar-aware interval. Can be specified using the unit name, such as `month`, or as a single unit quantity, such as `1M`.').optional(),
  extended_bounds: AggregationsExtendedBounds.describe('Enables extending the bounds of the histogram beyond the data itself.').optional(),
  hard_bounds: AggregationsExtendedBounds.describe('Limits the histogram to specified bounds.').optional(),
  field: z.lazy(() => Field).describe('The date field whose values are use to build a histogram.').optional(),
  fixed_interval: z.lazy(() => Duration).describe('Fixed intervals: a fixed number of SI units and never deviate, regardless of where they fall on the calendar.').optional(),
  format: z.string().describe('The date format used to format `key_as_string` in the response. If no `format` is specified, the first date format specified in the field mapping is used.').optional(),
  interval: z.lazy(() => Duration).optional(),
  min_doc_count: z.lazy(() => integer).describe('Only returns buckets that have `min_doc_count` number of documents. By default, all buckets between the first bucket that matches documents and the last one are returned.').optional(),
  missing: z.lazy(() => DateTime).describe('The value to apply to documents that do not have a value. By default, documents without a value are ignored.').optional(),
  offset: z.lazy(() => Duration).describe('Changes the start value of each bucket by the specified positive (`+`) or negative offset (`-`) duration.').optional(),
  order: AggregationsAggregateOrder.describe('The sort order of the returned buckets.').optional(),
  params: z.record(z.string(), z.any()).optional(),
  get script () { return z.union([Script, ScriptSource]).optional() },
  time_zone: z.lazy(() => TimeZone).describe('Time zone used for bucketing and rounding. Defaults to Coordinated Universal Time (UTC).').optional(),
  keyed: z.boolean().describe('Set to `true` to associate a unique string key with each bucket and return the ranges as a hash rather than an array.').optional()
}).meta({ id: 'AggregationsDateHistogramAggregation' })
export type AggregationsDateHistogramAggregation = z.infer<typeof AggregationsDateHistogramAggregation>
// ...
export const AggregationsTermsExclude = z.union([z.string(), z.array(z.string())]).meta({ id: 'AggregationsTermsExclude' })
export type AggregationsTermsExclude = z.infer<typeof AggregationsTermsExclude>

export const AggregationsTermsPartition = z.object({
  num_partitions: z.lazy(() => long).describe('The number of partitions.'),
  partition: z.lazy(() => long).describe('The partition number for this request.')
}).meta({ id: 'AggregationsTermsPartition' })
export type AggregationsTermsPartition = z.infer<typeof AggregationsTermsPartition>

export const AggregationsTermsInclude = z.union([z.string(), z.array(z.string()), AggregationsTermsPartition]).meta({ id: 'AggregationsTermsInclude' })
export type AggregationsTermsInclude = z.infer<typeof AggregationsTermsInclude>
// ...
export const AggregationsGeoTileGridAggregation = z.object({
  field: z.lazy(() => Field).describe('Field containing indexed `geo_point` or `geo_shape` values. If the field contains an array, `geotile_grid` aggregates all array values.').optional(),
  precision: z.lazy(() => GeoTilePrecision).describe('Integer zoom of the key used to define cells/buckets in the results. Values outside of the range [0,29] will be rejected.').optional(),
  shard_size: z.lazy(() => integer).describe('Allows for more accurate counting of the top cells returned in the final result the aggregation. Defaults to returning `max(10,(size x number-of-shards))` buckets from each shard.').optional(),
  size: z.lazy(() => integer).describe('The maximum number of buckets to return.').optional(),
  bounds: z.lazy(() => GeoBounds).describe('A bounding box to filter the geo-points or geo-shapes in each bucket.').optional()
}).meta({ id: 'AggregationsGeoTileGridAggregation' })
export type AggregationsGeoTileGridAggregation = z.infer<typeof AggregationsGeoTileGridAggregation>
// ...
export const AggregationsHistogramAggregation = z.object({
  extended_bounds: AggregationsExtendedBounds.describe('Enables extending the bounds of the histogram beyond the data itself.').optional(),
  hard_bounds: AggregationsExtendedBounds.describe('Limits the range of buckets in the histogram. It is particularly useful in the case of open data ranges that can result in a very large number of buckets.').optional(),
  field: z.lazy(() => Field).describe('The name of the field to aggregate on.').optional(),
  interval: z.lazy(() => double).describe('The interval for the buckets. Must be a positive decimal.').optional(),
  min_doc_count: z.lazy(() => integer).describe('Only returns buckets that have `min_doc_count` number of documents. By default, the response will fill gaps in the histogram with empty buckets.').optional(),
  missing: z.lazy(() => double).describe('The value to apply to documents that do not have a value. By default, documents without a value are ignored.').optional(),
  offset: z.lazy(() => double).describe('By default, the bucket keys start with 0 and then continue in even spaced steps of `interval`. The bucket boundaries can be shifted by using the `offset` option.').optional(),
  order: AggregationsAggregateOrder.describe('The sort order of the returned buckets. By default, the returned buckets are sorted by their key ascending.').optional(),
  get script () { return z.union([Script, ScriptSource]).optional() },
  format: z.string().optional(),
  keyed: z.boolean().describe('If `true`, returns buckets as a hash instead of an array, keyed by the bucket keys.').optional()
}).meta({ id: 'AggregationsHistogramAggregation' })
export type AggregationsHistogramAggregation = z.infer<typeof AggregationsHistogramAggregation>
// ...
export const AggregationsTermsAggregationCollectMode = z.enum(['depth_first', 'breadth_first']).meta({ id: 'AggregationsTermsAggregationCollectMode' })
export type AggregationsTermsAggregationCollectMode = z.infer<typeof AggregationsTermsAggregationCollectMode>
// ...
export const AggregationsTermsAggregationExecutionHint = z.enum(['map', 'global_ordinals', 'global_ordinals_hash', 'global_ordinals_low_cardinality']).meta({ id: 'AggregationsTermsAggregationExecutionHint' })
export type AggregationsTermsAggregationExecutionHint = z.infer<typeof AggregationsTermsAggregationExecutionHint>
// ...
export const AggregationsTermsAggregation = z.object({
  collect_mode: AggregationsTermsAggregationCollectMode.describe('Determines how child aggregations should be calculated: breadth-first or depth-first.').optional(),
  exclude: AggregationsTermsExclude.describe('Values to exclude. Accepts regular expressions and partitions.').optional(),
  execution_hint: AggregationsTermsAggregationExecutionHint.describe('Determines whether the aggregation will use field values directly or global ordinals.').optional(),
  field: z.lazy(() => Field).describe('The field from which to return terms.').optional(),
  include: AggregationsTermsInclude.describe('Values to include. Accepts regular expressions and partitions.').optional(),
  min_doc_count: z.lazy(() => integer).describe('Only return values that are found in more than `min_doc_count` hits.').optional(),
  missing: z.lazy(() => AggregationsMissing).describe('The value to apply to documents that do not have a value. By default, documents without a value are ignored.').optional(),
  missing_order: z.lazy(() => AggregationsMissingOrder).optional(),
  missing_bucket: z.boolean().optional(),
  value_type: z.string().describe('Coerced unmapped fields into the specified type.').optional(),
  order: AggregationsAggregateOrder.describe('Specifies the sort order of the buckets. Defaults to sorting by descending document count.').optional(),
  get script () { return z.union([Script, ScriptSource]).optional() },
  shard_min_doc_count: z.lazy(() => long).describe('Regulates the certainty a shard has if the term should actually be added to the candidate list or not with respect to the `min_doc_count`. Terms will only be considered if their local shard frequency within the set is higher than the `shard_min_doc_count`.').optional(),
  shard_size: z.lazy(() => integer).describe('The number of candidate terms produced by each shard. By default, `shard_size` will be automatically estimated based on the number of shards and the `size` parameter.').optional(),
  show_term_doc_count_error: z.boolean().describe('Set to `true` to return the `doc_count_error_upper_bound`, which is an upper bound to the error on the `doc_count` returned by each shard.').optional(),
  size: z.lazy(() => integer).describe('The number of buckets returned out of the overall terms list.').optional(),
  format: z.string().optional()
}).meta({ id: 'AggregationsTermsAggregation' })
export type AggregationsTermsAggregation = z.infer<typeof AggregationsTermsAggregation>
// ...
const AggregationsAggregationContainerCommonProps = z.object({
  aggregations: z.record(z.string(), z.lazy(() => AggregationsAggregationContainer)).describe('Sub-aggregations for this aggregation. Only applies to bucket aggregations.').optional(),
  aggs: z.record(z.string(), z.lazy(() => AggregationsAggregationContainer)).describe('Sub-aggregations for this aggregation. Only applies to bucket aggregations.').optional(),
  meta: z.lazy(() => Metadata).optional()
})
// ...
export const AggregationsAggregationContainer: z.ZodType<AggregationsAggregationContainerShape> = AggregationsAggregationContainerCommonProps.and(AggregationsAggregationContainerExclusiveProps).meta({ id: 'AggregationsAggregationContainer' })
export type AggregationsAggregationContainer = z.infer<typeof AggregationsAggregationContainer>